# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Jordan's 40th Trip Hub API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /expenses:
    get:
      operationId: listExpenses
      tags: [expenses]
      summary: List all expenses with shares
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Expense"
    post:
      operationId: createExpense
      tags: [expenses]
      summary: Create an expense with shares
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
  /expenses/{id}:
    patch:
      operationId: updateExpense
      tags: [expenses]
      summary: Update an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExpenseUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Expense"
    delete:
      operationId: deleteExpense
      tags: [expenses]
      summary: Delete an expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---- Balances ----
...
  /settlements:
    get:
      operationId: listSettlements
      tags: [settlements]
      summary: List all settlements
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Settlement"
    post:
      operationId: createSettlement
      tags: [settlements]
      summary: Record a settlement payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"
...
  /settlements/{id}:
    patch:
      operationId: updateSettlement
      tags: [settlements]
      summary: Mark a settlement as paid or pending
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettlementUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settlement"

  # ---- Guide ----
  /guide:
    get:
      operationId: listGuideEntries
      tags: [guide]
      summary: List guide entries, optionally filtered by type
      parameters:
        - name: type
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GuideEntry"
    post:
      operationId: createGuideEntry
      tags: [guide]
      summary: Create a guide entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuideEntryInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuideEntry"
  /guide/{id}:
    patch:
      operationId: updateGuideEntry
      tags: [guide]
      summary: Update a guide entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuideEntryUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GuideEntry"
    delete:
      operationId: deleteGuideEntry
      tags: [guide]
      summary: Delete a guide entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---- Contacts ----
  /contacts:
    get:
      operationId: listContacts
      tags: [contacts]
      summary: List all contacts
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Contact"
    post:
      operationId: createContact
      tags: [contacts]
      summary: Create a contact
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactInput"
      responses:
        "200":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
  /contacts/{id}:
    patch:
      operationId: updateContact
      tags: [contacts]
      summary: Update a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contact"
    delete:
      operationId: deleteContact
      tags: [contacts]
      summary: Delete a contact
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---- Settings ----
  /settings:
    get:
      operationId: getSettings
      tags: [settings]
      summary: Get all settings as key-value map
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsMap"
    patch:
      operationId: updateSettings
      tags: [settings]
      summary: Upsert one or more settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsMap"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkResponse"

  # ---- Itinerary ----
...
components:
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required:
        - ok

    # ---- Households ----
...
    ExpenseShare:
      type: object
      properties:
        householdId:  { type: number }
        shareAmount:  { type: number }
      required: [householdId, shareAmount]
    Expense:
      type: object
      properties:
        id:                     { type: number }
        date:                   { type: string }
        description:            { type: string }
        category:               { type: string }
        merchant:               { type: ["string", "null"] }
        paidByHouseholdId:      { type: ["number", "null"] }
        totalAmount:            { type: string }
        allocationMethod:       { type: string }
        participantHouseholdIds:
          type: array
          items: { type: number }
        receiptUrl:             { type: ["string", "null"] }
        notes:                  { type: ["string", "null"] }
        shares:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseShare"
      required: [id, date, description, category, totalAmount, allocationMethod]
    ExpenseInput:
      type: object
      properties:
        date:                   { type: string }
        description:            { type: string }
        category:               { type: string }
        merchant:               { type: ["string", "null"] }
        paidByHouseholdId:      { type: ["number", "null"] }
        totalAmount:            { type: string }
        allocationMethod:       { type: string }
        participantHouseholdIds:
          type: array
          items: { type: number }
        receiptUrl:             { type: ["string", "null"] }
        notes:                  { type: ["string", "null"] }
        shares:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseShare"
      required: [date, description, category, totalAmount, allocationMethod]
    ExpenseUpdate:
      type: object
      properties:
        date:                   { type: string }
        description:            { type: string }
        category:               { type: string }
        merchant:               { type: ["string", "null"] }
        paidByHouseholdId:      { type: ["number", "null"] }
        totalAmount:            { type: string }
        allocationMethod:       { type: string }
        participantHouseholdIds:
          type: array
          items: { type: number }
        receiptUrl:             { type: ["string", "null"] }
        notes:                  { type: ["string", "null"] }
        shares:
          type: array
          items:
            $ref: "#/components/schemas/ExpenseShare"

    # ---- Balances ----
...
    Settlement:
      type: object
      properties:
        id:               { type: number }
        fromHouseholdId:  { type: number }
        toHouseholdId:    { type: number }
        amount:           { type: string }
        status:           { type: string }
        createdAt:        { type: ["string", "null"] }
        paidAt:           { type: ["string", "null"] }
      required: [id, fromHouseholdId, toHouseholdId, amount, status]
    SettlementInput:
      type: object
      properties:
        fromHouseholdId:  { type: number }
        toHouseholdId:    { type: number }
        amount:           { type: string }
        status:           { type: string }
      required: [fromHouseholdId, toHouseholdId, amount]
    SettlementUpdate:
      type: object
      properties:
        status:           { type: string }
...
    GuideEntry:
      type: object
      properties:
        id:                 { type: number }
        type:               { type: string }
        name:               { type: string }
        subtitle:           { type: ["string", "null"] }
        priceTier:          { type: ["string", "null"] }
        distanceFromVilla:  { type: ["string", "null"] }
        notes:              { type: ["string", "null"] }
        photoUrl:           { type: ["string", "null"] }
        mapUrl:             { type: ["string", "null"] }
        reservationStatus:  { type: ["string", "null"] }
        websiteUrl:         { type: ["string", "null"] }
        favorited:          { type: ["boolean", "null"] }
        sortOrder:          { type: ["number", "null"] }
      required: [id, type, name]
    GuideEntryInput:
      type: object
      properties:
        type:               { type: string }
        name:               { type: string }
        subtitle:           { type: ["string", "null"] }
        priceTier:          { type: ["string", "null"] }
        distanceFromVilla:  { type: ["string", "null"] }
        notes:              { type: ["string", "null"] }
        photoUrl:           { type: ["string", "null"] }
        mapUrl:             { type: ["string", "null"] }
        reservationStatus:  { type: ["string", "null"] }
        websiteUrl:         { type: ["string", "null"] }
        favorited:          { type: boolean }
        sortOrder:          { type: number }
      required: [type, name]
    GuideEntryUpdate:
      type: object
      properties:
        type:               { type: string }
        name:               { type: string }
        subtitle:           { type: ["string", "null"] }
        priceTier:          { type: ["string", "null"] }
        distanceFromVilla:  { type: ["string", "null"] }
        notes:              { type: ["string", "null"] }
        photoUrl:           { type: ["string", "null"] }
        mapUrl:             { type: ["string", "null"] }
        reservationStatus:  { type: ["string", "null"] }
        websiteUrl:         { type: ["string", "null"] }
        favorited:          { type: boolean }
        sortOrder:          { type: number }

    # ---- Contacts ----
    Contact:
      type: object
      properties:
        id:        { type: number }
        category:  { type: string }
        name:      { type: string }
        phone:     { type: ["string", "null"] }
        address:   { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        sortOrder: { type: ["number", "null"] }
      required: [id, category, name]
    ContactInput:
      type: object
      properties:
        category:  { type: string }
        name:      { type: string }
        phone:     { type: ["string", "null"] }
        address:   { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        sortOrder: { type: number }
      required: [category, name]
    ContactUpdate:
      type: object
      properties:
        category:  { type: string }
        name:      { type: string }
        phone:     { type: ["string", "null"] }
        address:   { type: ["string", "null"] }
        notes:     { type: ["string", "null"] }
        sortOrder: { type: number }

    # ---- Settings ----
    SettingsMap:
      type: object
      additionalProperties:
        type: ["string", "null"]

    # ---- Itinerary ----
...