# lib/api-zod/src/generated/types/contact.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface Contact {
  id: number;
  category: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  sortOrder?: number | null;
}