# lib/api-zod/src/generated/types/expenseShare.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface ExpenseShare {
  householdId: number;
  shareAmount: number;
}