# lib/api-zod/src/generated/types/expenseUpdate.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ExpenseShare } from './expenseShare';

export interface ExpenseUpdate {
  date?: string;
  description?: string;
  category?: string;
  /** @nullable */
  merchant?: string | null;
  /** @nullable */
  paidByHouseholdId?: number | null;
  totalAmount?: string;
  allocationMethod?: string;
  participantHouseholdIds?: number[];
  /** @nullable */
  receiptUrl?: string | null;
  /** @nullable */
  notes?: string | null;
  shares?: ExpenseShare[];
}