# lib/api-zod/src/generated/types/guideEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Jordan's 40th Trip Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface GuideEntry {
  id: number;
  type: string;
  name: string;
  /** @nullable */
  subtitle?: string | null;
  /** @nullable */
  priceTier?: string | null;
  /** @nullable */
  distanceFromVilla?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** @nullable */
  mapUrl?: string | null;
  /** @nullable */
  reservationStatus?: string | null;
  /** @nullable */
  websiteUrl?: string | null;
  /** @nullable */
  favorited?: boolean | null;
  /** @nullable */
  sortOrder?: number | null;
}